move header.js to the components folder
